extracted locals test
