+++ test/integration_test.go
@@ -116,13 +116,6 @@ const (
 	testFixtureAutoRetryConfigurableRetriesError2                 = "fixture-auto-retry/configurable-retries-incorrect-sleep-interval"
 	testFixtureAwsProviderPatch                                   = "fixture-aws-provider-patch"
 	testFixtureInputs                                             = "fixture-inputs"
-	testFixtureLocalsErrorUndefinedLocal                          = "fixture-locals-errors/undefined-local"
-	testFixtureLocalsErrorUndefinedLocalButInput                  = "fixture-locals-errors/undefined-local-but-input"
-	testFixtureLocalsCanonical                                    = "fixture-locals/canonical"
-	testFixtureLocalsInInclude                                    = "fixture-locals/local-in-include"
-	TestFixtureLocalRunOnce                                       = "fixture-locals/run-once"
-	testFixtureLocalRunMultiple                                   = "fixture-locals/run-multiple"
-	testFixtureLocalsInIncludeChildRelPath                        = "qa/my-app"
 	testFixtureNoColor                                            = "fixture-no-color"
 	testFixtureReadConfig                                         = "fixture-read-config"
 	testFixtureReadIamRole                                        = "fixture-read-config/iam_role_in_file"
@@ -146,7 +139,6 @@ const (
 	testFixtureIncludeParent                                      = "fixture-include-parent"
 	testFixtureAutoInit                                           = "fixture-download/init-on-source-change"
 	testFixtureDisjoint                                           = "fixture-stack/disjoint"
-	testFixtureBrokenLocals                                       = "fixture-broken-locals"
 	testFixtureBrokenDependency                                   = "fixture-broken-dependency"
 	testFixtureRenderJsonMetadata                                 = "fixture-render-json-metadata"
 	testFixtureRenderJsonMockOutputs                              = "fixture-render-json-mock-outputs"
@@ -1926,84 +1918,6 @@ func TestNoAutoInit(t *testing.T) {
 	logBufferContentsLineByLine(t, stderr, "no force apply stderr")
 	require.Error(t, err)
 	assert.Contains(t, stderr.String(), "This module is not yet installed.")
-}
-
-func TestLocalsParsing(t *testing.T) {
-	t.Parallel()
-
-	cleanupTerraformFolder(t, testFixtureLocalsCanonical)
-
-	runTerragrunt(t, "terragrunt apply -auto-approve --terragrunt-non-interactive --terragrunt-working-dir "+testFixtureLocalsCanonical)
-
-	stdout := bytes.Buffer{}
-	stderr := bytes.Buffer{}
-
-	err := runTerragruntCommand(t, "terragrunt output -no-color -json --terragrunt-non-interactive --terragrunt-working-dir "+testFixtureLocalsCanonical, &stdout, &stderr)
-	require.NoError(t, err)
-
-	outputs := map[string]TerraformOutput{}
-	require.NoError(t, json.Unmarshal(stdout.Bytes(), &outputs))
-
-	assert.Equal(t, "Hello world\n", outputs["data"].Value)
-	assert.InEpsilon(t, 42.0, outputs["answer"].Value, 0.0000000001)
-}
-
-func TestLocalsInInclude(t *testing.T) {
-	t.Parallel()
-
-	cleanupTerraformFolder(t, testFixtureLocalsInInclude)
-	tmpEnvPath := copyEnvironment(t, testFixtureLocalsInInclude)
-	childPath := filepath.Join(tmpEnvPath, testFixtureLocalsInInclude, testFixtureLocalsInIncludeChildRelPath)
-	runTerragrunt(t, "terragrunt apply -auto-approve -no-color --terragrunt-non-interactive --terragrunt-working-dir "+childPath)
-
-	// Check the outputs of the dir functions referenced in locals to make sure they return what is expected
-	stdout := bytes.Buffer{}
-	stderr := bytes.Buffer{}
-
-	require.NoError(
-		t,
-		runTerragruntCommand(t, "terragrunt output -no-color -json --terragrunt-non-interactive --terragrunt-working-dir "+childPath, &stdout, &stderr),
-	)
-
-	outputs := map[string]TerraformOutput{}
-	require.NoError(t, json.Unmarshal(stdout.Bytes(), &outputs))
-
-	assert.Equal(
-		t,
-		filepath.Join(tmpEnvPath, testFixtureLocalsInInclude),
-		outputs["parent_terragrunt_dir"].Value,
-	)
-	assert.Equal(
-		t,
-		childPath,
-		outputs["terragrunt_dir"].Value,
-	)
-	assert.Equal(
-		t,
-		"apply",
-		outputs["terraform_command"].Value,
-	)
-	assert.Equal(
-		t,
-		"[\"apply\",\"-auto-approve\",\"-no-color\"]",
-		outputs["terraform_cli_args"].Value,
-	)
-}
-
-func TestUndefinedLocalsReferenceBreaks(t *testing.T) {
-	t.Parallel()
-
-	cleanupTerraformFolder(t, testFixtureLocalsErrorUndefinedLocal)
-	err := runTerragruntCommand(t, "terragrunt apply -auto-approve --terragrunt-non-interactive --terragrunt-working-dir "+testFixtureLocalsErrorUndefinedLocal, os.Stdout, os.Stderr)
-	require.Error(t, err)
-}
-
-func TestUndefinedLocalsReferenceToInputsBreaks(t *testing.T) {
-	t.Parallel()
-
-	cleanupTerraformFolder(t, testFixtureLocalsErrorUndefinedLocalButInput)
-	err := runTerragruntCommand(t, "terragrunt apply -auto-approve --terragrunt-non-interactive --terragrunt-working-dir "+testFixtureLocalsErrorUndefinedLocalButInput, os.Stdout, os.Stderr)
-	require.Error(t, err)
 }
 
 type TerraformOutput struct {
@@ -4401,21 +4315,6 @@ func TestTerragruntVersionConstraintsPartialParse(t *testing.T) {
 	assert.True(t, ok)
 }
 
-func TestLogFailedLocalsEvaluation(t *testing.T) {
-	t.Parallel()
-
-	var (
-		stdout bytes.Buffer
-		stderr bytes.Buffer
-	)
-
-	err := runTerragruntCommand(t, fmt.Sprintf("terragrunt apply -auto-approve --terragrunt-non-interactive --terragrunt-working-dir %s --terragrunt-log-level debug", testFixtureBrokenLocals), &stdout, &stderr)
-	require.Error(t, err)
-
-	output := stderr.String()
-	assert.Contains(t, output, "Encountered error while evaluating locals in file terragrunt.hcl")
-}
-
 func TestLogFailingDependencies(t *testing.T) {
 	t.Parallel()
 
@@ -5131,47 +5030,6 @@ func TestTerragruntRunAllCommandPrompt(t *testing.T) {
 	logBufferContentsLineByLine(t, stderr, "stderr")
 	assert.Contains(t, stderr.String(), "Are you sure you want to run 'terragrunt apply' in each folder of the stack described above? (y/n)")
 	require.Error(t, err)
-}
-
-func TestTerragruntLocalRunOnce(t *testing.T) {
-	t.Parallel()
-
-	cleanupTerraformFolder(t, TestFixtureLocalRunOnce)
-	stdout := bytes.Buffer{}
-	stderr := bytes.Buffer{}
-
-	err := runTerragruntCommand(t, "terragrunt init --terragrunt-working-dir "+TestFixtureLocalRunOnce, &stdout, &stderr)
-	require.Error(t, err)
-
-	errout := stdout.String()
-
-	assert.Equal(t, 1, strings.Count(errout, "foo"))
-}
-
-func TestTerragruntInitRunCmd(t *testing.T) {
-	t.Parallel()
-
-	cleanupTerraformFolder(t, testFixtureLocalRunMultiple)
-	stdout := bytes.Buffer{}
-	stderr := bytes.Buffer{}
-
-	err := runTerragruntCommand(t, "terragrunt init --terragrunt-working-dir "+testFixtureLocalRunMultiple, &stdout, &stderr)
-	require.Error(t, err)
-
-	errout := stdout.String()
-
-	// Check for cached values between locals and inputs sections
-	assert.Equal(t, 1, strings.Count(errout, "potato"))
-	assert.Equal(t, 1, strings.Count(errout, "carrot"))
-	assert.Equal(t, 1, strings.Count(errout, "bar"))
-	assert.Equal(t, 1, strings.Count(errout, "foo"))
-
-	assert.Equal(t, 1, strings.Count(errout, "input_variable"))
-
-	// Commands executed multiple times because of different arguments
-	assert.Equal(t, 4, strings.Count(errout, "uuid"))
-	assert.Equal(t, 6, strings.Count(errout, "random_arg"))
-	assert.Equal(t, 4, strings.Count(errout, "another_arg"))
 }
 
 func TestShowWarningWithDependentModulesBeforeDestroy(t *testing.T) {
